requirements.txt
Django>=5.1.5
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
Pillow>=10.2.0  # For image handling
django-filter>=23.5  # For filtering capabilities
